feat(payment-integration) enhance payment processing with subscription validation and update transaction handling

+++ src/main/kotlin/com/exe/vocafy_BE/model/entity/SubscriptionTransaction.kt
@@ -32,7 +32,7 @@ class SubscriptionTransaction(
     val paymentMethod: PaymentMethod,
 
     @Column(name = "amount", nullable = false)
-    val amount: Int,
+    val amount: Long,
 
     @Enumerated(EnumType.STRING)
     @Column(name = "status", nullable = false, length = 20)
